+++ Extensions/IISWebAppDeploy/Src/Tasks/IISWebAppDeploy/DeployIISWebApp.ps1
@@ -54,6 +54,8 @@ function Get-ScriptToRun
     )
 
     $msDeployScript = Get-Content  ./MsDeployOnTargetMachines.ps1 | Out-String
+    $overRideParams = Escape-DoubleQuotes -str $overRideParams
+
     $invokeMain = "Execute-Main -WebDeployPackage `"$webDeployPackage`" -WebDeployParamFile `"$webDeployParamFile`" -OverRideParams `"$overRideParams`""
 
     Write-Verbose "Executing main funnction in MsDeployOnTargetMachines : $invokeMain"
@@ -70,7 +72,8 @@ function Run-RemoteDeployment
         [string]$adminPassword,
         [string]$winrmProtocol,
         [string]$testCertificate,
-        [string]$deployInParallel
+        [string]$deployInParallel,
+        [string]$webDeployPackage
     )
 
     Write-Host "Starting deployment of IIS Web Deploy Package : $webDeployPackage"
@@ -79,7 +82,7 @@ function Run-RemoteDeployment
 
     if(-not [string]::IsNullOrEmpty($errorMessage))
     {
-        $helpMessage = "For more info please refer to http://aka.ms/iisextnreadme)"
+        $helpMessage = "For more info please refer to http://aka.ms/iisextnreadme"
         Write-Error "$errorMessage`n$helpMessage"
         return
     }
@@ -114,7 +117,6 @@ function Main
 
     Trim-Inputs -package ([ref]$webDeployPackage) -paramFile ([ref]$webDeployParamFile) -siteName ([ref]$websiteName) -adminUser ([ref]$adminUserName)
     $overRideParams = Compute-MsDeploy-SetParams -websiteName $websiteName -overRideParams $overRideParams
-    $overRideParams = Escape-DoubleQuotes -str $overRideParams
     $script = Get-ScriptToRun -webDeployPackage $webDeployPackage -webDeployParamFile $webDeployParamFile -overRideParams $overRideParams
-    Run-RemoteDeployment -machinesList $machinesList -scriptToRun $script -adminUserName $adminUserName -adminPassword $adminPassword -winrmProtocol $winrmProtocol -testCertificate $testCertificate -deployInParallel $deployInParallel
+    Run-RemoteDeployment -machinesList $machinesList -scriptToRun $script -adminUserName $adminUserName -adminPassword $adminPassword -winrmProtocol $winrmProtocol -testCertificate $testCertificate -deployInParallel $deployInParallel -webDeployPackage $webDeployPackage
 }