+++ Extensions/IISWebAppDeploy/Src/Tasks/IISWebAppMgmt/ManageIISWebApp.ps1
@@ -129,7 +129,7 @@ function Run-RemoteDeployment
 
     if(-not [string]::IsNullOrEmpty($errorMessage))
     {
-        $helpMessage = "For more info please refer to http://aka.ms/iisextnreadme)"
+        $helpMessage = "For more info please refer to http://aka.ms/iisextnreadme"
         Write-Error "$errorMessage`n$helpMessage"
         return
     }
@@ -200,7 +200,7 @@ function Main
     Write-Verbose "appCmdCommands = $appCmdCommands"
     Write-Verbose "deployInParallel = $deployInParallel"
 
-    Trim-Inputs -package -siteName ([ref]$websiteName) -physicalPath ([ref]$websitePhysicalPath)  -poolName ([ref]$appPoolName) -websitePathAuthuser ([ref]$websiteAuthUserName) -appPoolUser ([ref]$appPoolUsername) -adminUser ([ref]$adminUserName)
+    Trim-Inputs -siteName ([ref]$websiteName) -physicalPath ([ref]$websitePhysicalPath)  -poolName ([ref]$appPoolName) -websitePathAuthuser ([ref]$websiteAuthUserName) -appPoolUser ([ref]$appPoolUsername) -adminUser ([ref]$adminUserName)
 
     Validate-Inputs -createWebsite $createWebsite -websiteName $websiteName -createAppPool $createAppPool -appPoolName $appPoolName